+++ tests/unit/midi_out.cpp
@@ -1,4 +1,4 @@
-#include "include_catch.hpp"
+#include "../include_catch.hpp"
 
 #include <libremidi/libremidi.hpp>
